database: Move verifier functions to their own file.
